Re-format data structure and add uniform neutralizing background
The position and velocity data structures are now MATLAB arrays that have to be unpacked by species, instead of cell arrays.  This helps with outputting these parameters  with an arbitrary number of species.

The determination of rho in function QtoGrid now includes a few lines to ensure the average rho is zero, which acts as a neutralizing uniform density.