PIC fixed and overhaul of Species class
The CIC formulae for getting rho were missing a factor of 1/dx method, which was fixed.  Also, the mean velocity was being calculated incorrectly, which was fixed.

The Species class was adjusted so the the user sets the plasma frequency, wp, and the charge-to-mass ratio, qm., which is consistent with Birdsall and Langdon.  Previously, the user would set the charge, q, and the mass, m.  Species was also enhanced to have a method that sets default zero velocity arrays and uniform distribution.